[frontend] When putting number widgets in the middle of a dashboard, it breaks horizontal alignement (#14856)
