+++ opencti-platform/opencti-front/lang/front/de.json
@@ -395,6 +395,7 @@
   "Audit logs": "Audit-Protokolle",
   "Audits are not supported in public dashboards": "Audits werden in öffentlichen Dashboards nicht unterstützt",
   "Audits list": "Liste der Audits",
+  "Audits number": "Anzahl der Audits",
   "Authentication": "Authentifizierung",
   "Authentication (AU)": "Authentifizierung (AU)",
   "Authentication | Security | Settings": "Authentifizierung | Sicherheit | Einstellungen",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Geschaffene Beziehungen",
   "Relationships distribution": "Relationships-Verteilung",
   "Relationships list": "Liste der Beziehungen",
+  "Relationships number": "Nummer der Beziehungen",
   "Relationships timeline": "Beziehungen Zeitleiste",
   "Relationships view": "Ansicht \"Beziehungen",
   "Relative time": "Relative Zeit",
