+++ opencti-platform/opencti-front/lang/front/en.json
@@ -395,6 +395,7 @@
   "Audit logs": "Audit logs",
   "Audits are not supported in public dashboards": "Audits are not supported in public dashboards",
   "Audits list": "Audits list",
+  "Audits number": "Audits number",
   "Authentication": "Authentication",
   "Authentication (AU)": "Authentication (AU)",
   "Authentication | Security | Settings": "Authentication | Security | Settings",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Relationships created",
   "Relationships distribution": "Relationships distribution",
   "Relationships list": "Relationships list",
+  "Relationships number": "Relationships number",
   "Relationships timeline": "Relationships timeline",
   "Relationships view": "Relationships view",
   "Relative time": "Relative time",
