+++ opencti-platform/opencti-front/lang/front/es.json
@@ -395,6 +395,7 @@
   "Audit logs": "Registros de auditoría",
   "Audits are not supported in public dashboards": "Los cuadros de mando públicos no admiten auditorías",
   "Audits list": "Lista de auditorías",
+  "Audits number": "Número de auditorías",
   "Authentication": "Autenticación",
   "Authentication (AU)": "Autenticación (AU)",
   "Authentication | Security | Settings": "Autenticación | Seguridad | Configuración",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Relaciones creadas",
   "Relationships distribution": "Distribución de relaciones",
   "Relationships list": "Lista de relaciones",
+  "Relationships number": "Número de relaciones",
   "Relationships timeline": "Cronología de relaciones",
   "Relationships view": "Vista de relaciones",
   "Relative time": "Tiempo relativo",
