+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -395,6 +395,7 @@
   "Audit logs": "Journaux d'audit",
   "Audits are not supported in public dashboards": "Les audits ne sont pas pris en charge dans les tableaux de bord publics",
   "Audits list": "Liste des audits",
+  "Audits number": "Nombre d'audits",
   "Authentication": "Authentification",
   "Authentication (AU)": "Authentification (AU)",
   "Authentication | Security | Settings": "Authentification | Sécurité | Paramètres",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Relations créées",
   "Relationships distribution": "Distribution des relations",
   "Relationships list": "Liste des relations",
+  "Relationships number": "Numéro de relations",
   "Relationships timeline": "Chronologie des relations",
   "Relationships view": "Vue relations",
   "Relative time": "Temps relatif",
