+++ opencti-platform/opencti-front/lang/front/it.json
@@ -395,6 +395,7 @@
   "Audit logs": "Log di audit",
   "Audits are not supported in public dashboards": "Gli audit non sono supportati nelle dashboard pubbliche",
   "Audits list": "Elenco degli audit",
+  "Audits number": "Numero di audit",
   "Authentication": "Autenticazione",
   "Authentication (AU)": "Autenticazione (AU)",
   "Authentication | Security | Settings": "Autenticazione | Sicurezza | Impostazioni",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Relazioni create",
   "Relationships distribution": "Distribuzione delle relazioni",
   "Relationships list": "Elenco delle relazioni",
+  "Relationships number": "Numero di relazioni",
   "Relationships timeline": "Cronologia delle relazioni",
   "Relationships view": "Vista delle relazioni",
   "Relative time": "Tempo relativo",
