+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -395,6 +395,7 @@
   "Audit logs": "監査ログ",
   "Audits are not supported in public dashboards": "パブリックダッシュボードでは監査はサポートされていません。",
   "Audits list": "監査リスト",
+  "Audits number": "監査数",
   "Authentication": "認証",
   "Authentication (AU)": "認証 (AU)",
   "Authentication | Security | Settings": "認証｜セキュリティ｜設定",
@@ -3508,6 +3509,7 @@
   "Relationships created": "作成された関係",
   "Relationships distribution": "関係ディストリビューション",
   "Relationships list": "関係リスト",
+  "Relationships number": "関係番号",
   "Relationships timeline": "関係タイムライン",
   "Relationships view": "関係ビュー",
   "Relative time": "相対時間",
