+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -395,6 +395,7 @@
   "Audit logs": "감사 로그",
   "Audits are not supported in public dashboards": "공개 대시보드에서는 감사가 지원되지 않습니다.",
   "Audits list": "감사 목록",
+  "Audits number": "감사 번호",
   "Authentication": "인증",
   "Authentication (AU)": "인증(AU)",
   "Authentication | Security | Settings": "인증 | 보안 | 설정",
@@ -3508,6 +3509,7 @@
   "Relationships created": "생성된 관계",
   "Relationships distribution": "관계 분포",
   "Relationships list": "관계 목록",
+  "Relationships number": "관계 번호",
   "Relationships timeline": "관계 타임라인",
   "Relationships view": "관계 보기",
   "Relative time": "상대 시간",
