+++ opencti-platform/opencti-front/lang/front/ru.json
@@ -395,6 +395,7 @@
   "Audit logs": "Журналы аудита",
   "Audits are not supported in public dashboards": "Аудиты не поддерживаются в публичных информационных панелях",
   "Audits list": "Список аудитов",
+  "Audits number": "Количество аудитов",
   "Authentication": "Аутентификация",
   "Authentication (AU)": "Аутентификация (AU)",
   "Authentication | Security | Settings": "Аутентификация | Безопасность | Настройки",
@@ -3508,6 +3509,7 @@
   "Relationships created": "Созданные связи",
   "Relationships distribution": "Распределение связей",
   "Relationships list": "Список связей",
+  "Relationships number": "Количество взаимоотношений",
   "Relationships timeline": "Хронология связей",
   "Relationships view": "Просмотр связей",
   "Relative time": "Относительное время",
