+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -395,6 +395,7 @@
   "Audit logs": "审计日志",
   "Audits are not supported in public dashboards": "公共仪表盘不支持审计",
   "Audits list": "审核列表",
+  "Audits number": "审计编号",
   "Authentication": "验证",
   "Authentication (AU)": "验证 (AU)",
   "Authentication | Security | Settings": "身份验证 | 安全 | 设置",
@@ -3508,6 +3509,7 @@
   "Relationships created": "建立的关系",
   "Relationships distribution": "关系分布",
   "Relationships list": "关系列表",
+  "Relationships number": "关系编号",
   "Relationships timeline": "关系时间表",
   "Relationships view": "关系视图",
   "Relative time": "相对时间",
