+++ opencti-platform/opencti-front/src/components/common/card/CardNumber.tsx
@@ -1,84 +1,12 @@
-import { ReactNode } from 'react';
-import { useTheme } from '@mui/styles';
-import { Stack, Typography } from '@mui/material';
-import { Theme } from '../../Theme';
-import { useFormatter } from '../../i18n';
-import NumberDifference from '../../NumberDifference';
-import ItemIcon from '../../ItemIcon';
 import Card from './Card';
+import WidgetNumber, { WidgetNumberProps } from '../../dashboard/WidgetNumber';
 
-interface CardNumberProps {
-  label: string;
-  value: number;
-  diffLabel?: string;
-  diffValue?: number;
-  entityType?: string;
-  icon?: ReactNode;
-  action?: ReactNode;
-}
+type CardNumberProps = WidgetNumberProps;
 
-const CardNumber = ({
-  label,
-  value,
-  diffLabel,
-  diffValue,
-  entityType,
-  icon,
-  action,
-}: CardNumberProps) => {
-  const { n } = useFormatter();
-  const theme = useTheme<Theme>();
-
-  const valueStyle = {
-    fontSize: 32,
-    lineHeight: 1,
-    fontWeight: 600,
-  };
-
+const CardNumber = (props: CardNumberProps) => {
   return (
     <Card sx={{ paddingY: 2 }}>
-      <Stack height="100%" justifyContent="space-between">
-        <Stack direction="row" alignItems="start">
-          <Stack direction="row" alignItems="start" gap={1} flex={1}>
-            <Typography
-              color={theme.palette.text.light}
-              variant="body2"
-              gutterBottom
-            >
-              {label}
-            </Typography>
-            {diffValue !== undefined && diffLabel && (
-              <NumberDifference
-                value={diffValue}
-                description={diffLabel}
-              />
-            )}
-          </Stack>
-          {action}
-        </Stack>
-
-        <Stack
-          direction="row"
-          justifyContent="space-between"
-          alignItems="center"
-        >
-          <div
-            data-testid={`card-number-${label}`}
-            style={valueStyle}
-          >
-            {n(value)}
-          </div>
-          {entityType && (
-            <ItemIcon
-              type={entityType}
-              size="large"
-              color={theme.palette.text.secondary}
-              style={{ opacity: 0.35 }}
-            />
-          )}
-          {icon}
-        </Stack>
-      </Stack>
+      <WidgetNumber {...props} />
     </Card>
   );
 };
