+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectsNumber.jsx
@@ -1,5 +1,4 @@
 import { graphql } from 'react-relay';
-import CardNumber from '@common/card/CardNumber';
 import { QueryRenderer } from '../../../../relay/environment';
 import { useFormatter } from '../../../../components/i18n';
 import { dayAgo } from '../../../../utils/Time';
@@ -8,30 +7,31 @@ import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
 import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
 import useEntityTranslation from '../../../../utils/hooks/useEntityTranslation';
+import WidgetNumber from '../../../../components/dashboard/WidgetNumber';
 
 const stixCoreObjectsNumberNumberQuery = graphql`
-  query StixCoreObjectsNumberNumberSeriesQuery(
-    $dateAttribute: String
-    $types: [String]
-    $startDate: DateTime
-    $endDate: DateTime
-    $onlyInferred: Boolean
-    $filters: FilterGroup
-    $search: String
-  ) {
-    stixCoreObjectsNumber(
-      dateAttribute: $dateAttribute
-      types: $types
-      startDate: $startDate
-      endDate: $endDate
-      onlyInferred: $onlyInferred
-      filters: $filters
-      search: $search
+    query StixCoreObjectsNumberNumberSeriesQuery(
+        $dateAttribute: String
+        $types: [String]
+        $startDate: DateTime
+        $endDate: DateTime
+        $onlyInferred: Boolean
+        $filters: FilterGroup
+        $search: String
     ) {
-      total
-      count
+        stixCoreObjectsNumber(
+            dateAttribute: $dateAttribute
+            types: $types
+            startDate: $startDate
+            endDate: $endDate
+            onlyInferred: $onlyInferred
+            filters: $filters
+            search: $search
+        ) {
+            total
+            count
+        }
     }
-  }
 `;
 
 const StixCoreObjectsNumber = ({
@@ -41,6 +41,8 @@ const StixCoreObjectsNumber = ({
   parameters = {},
   entityType,
   popover,
+  variant,
+  height,
 }) => {
   const { t_i18n } = useFormatter();
   const { translateEntityType } = useEntityTranslation();
@@ -59,43 +61,42 @@ const StixCoreObjectsNumber = ({
   );
 
   return (
-    <QueryRenderer
-      query={stixCoreObjectsNumberNumberQuery}
-      variables={{
-        types: dataSelectionTypes,
-        dateAttribute,
-        filters,
-        startDate,
-        endDate: dayAgo(),
-      }}
-      render={({ props }) => {
-        if (props && props.stixCoreObjectsNumber) {
-          const { total, count } = props.stixCoreObjectsNumber;
-          return (
-            <CardNumber
-              entityType={entityType}
-              label={translatedTitle}
-              value={total}
-              diffLabel={t_i18n('24 hours')}
-              diffValue={total - count}
-              action={popover}
-            />
-          );
-        }
-        if (props) {
-          return (
-            <WidgetContainer title={title}>
-              <WidgetNoData />
-            </WidgetContainer>
-          );
-        }
-        return (
-          <WidgetContainer>
-            <Loader variant={LoaderVariant.inElement} />
-          </WidgetContainer>
-        );
-      }}
-    />
+    <WidgetContainer
+      padding="medium"
+      height={height}
+      title={t_i18n('Entities number')}
+      variant={variant}
+      action={popover}
+    >
+      <QueryRenderer
+        query={stixCoreObjectsNumberNumberQuery}
+        variables={{
+          types: dataSelectionTypes,
+          dateAttribute,
+          filters,
+          startDate,
+          endDate: dayAgo(),
+        }}
+        render={({ props }) => {
+          if (props && props.stixCoreObjectsNumber) {
+            const { total, count } = props.stixCoreObjectsNumber;
+            return (
+              <WidgetNumber
+                entityType={entityType}
+                label={translatedTitle}
+                value={total}
+                diffLabel={t_i18n('24 hours')}
+                diffValue={total - count}
+              />
+            );
+          }
+          if (props) {
+            return <WidgetNoData />;
+          }
+          return <Loader variant={LoaderVariant.inElement} />;
+        }}
+      />
+    </WidgetContainer>
   );
 };
